% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try-lavaanList.R
\name{compareFitProp}
\alias{compareFitProp}
\title{Compare Fitting Propensities}
\usage{
compareFitProp(
  ...,
  sameReps = TRUE,
  omit.reps = c("no.conv", "no.se"),
  fit.measures = "srmr",
  lower.tail = TRUE,
  probs = seq(0, 1, 0.1),
  type = 7,
  NML = FALSE,
  UIF = FALSE,
  conf = FALSE,
  nd = 3L
)
}
\arguments{
\item{...}{objects of class \code{\linkS4class{FitProp}}}

\item{sameReps}{\code{logical} indicating whether the whether to use only
replications in which all selected models converged}

\item{omit.reps}{\code{character} vector of conditions for removing
replications.  \code{"no.conv"} removes replications for which the model
did not converge.  \code{"no.se"} removes replications for which
\emph{SE}s could not be calculated.  \code{"no.hpd"} removes replications
with improper solutions (a nonpositive definite matrix).  \code{integer}s
can also be included in the vector, indicating particular replications
to remove.}

\item{fit.measures}{see \code{\link[lavaan]{fitMeasures}()}}

\item{lower.tail}{\code{logical} indicating whether lower values imply
better fit (default: \code{TRUE}, for "badness of fit" indices) or whether
higher values (\code{FALSE}) imply better ("goodness of") fit.}

\item{probs, type}{see \code{\link[stats]{quantile}}}

\item{NML, UIF}{\code{logical} indicating whether to calculate comparative
fit indices that adjust for a model's fitting propensity}

\item{conf}{confidence level for interval estimates of standardized
differences in central tendency. \code{FALSE} indicates to exclude them.}

\item{nd}{\code{integer}. Number of decimal places rounded in printed output}
}
\value{
For any models in \code{...}, quantiles are provided from the empirical
distribution(s) of their \code{fit.measures} to random data, along with
pairwise differences between distributions. Summary statistics for model
comparison (e.g., UIF) are also available.
}
\description{
Compare the empirical distributions of models' fitting propensity, or their
fit to the data adjusted for fitting propensity
}
\examples{
## See ?fitprop help page for examples

}
\references{
Botha, J. D., Shapiro, A., \& Steiger, J .H. (1988). Uniform indices-of-fit
for factor analysis models. \emph{Multivariate Behavioral Research, 23}(4),
 443--450. \doi{10.1207/s15327906mbr2304_2}

Rissanen, J. (2001). Strong optimality of the normalized ML models as
universal codes and information in data.
\emph{IEEE Transactions on Information Theory, 47}(5), 1712--1717.
\doi{10.1109/18.930912}
}
\author{
Terrence D. Jorgensen (University of Amsterdam; \email{TJorgensen314@gmail.com})
}
