% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try-lavaanList.R
\docType{class}
\name{FitProp-class}
\alias{FitProp-class}
\alias{show,FitProp-method}
\alias{summary,FitProp-method}
\title{Class for Fitting a Model to Random Data Patterns}
\usage{
\S4method{show}{FitProp}(object)

\S4method{summary}{FitProp}(
  object,
  ...,
  fit.measures = "srmr",
  lower.tail = TRUE,
  NML = FALSE,
  UIF = FALSE,
  probs = seq(0, 1, 0.1),
  type = 7,
  omit.reps = c("no.conv", "no.se")
)
}
\arguments{
\item{object}{an object of class \code{FitProp}}

\item{...}{arguments passed to/from other methods (not used)}

\item{fit.measures}{see \code{\link[lavaan]{fitMeasures}()}}

\item{lower.tail}{\code{logical} indicating whether lower values imply
better fit (default: \code{TRUE}, for "badness of fit" indices) or whether
higher values (\code{FALSE}) imply better ("goodness of") fit.}

\item{NML, UIF}{\code{logical} indicating whether to calculate comparative
fit indices that adjust for a model's fitting propensity}

\item{probs, type}{see \code{\link[stats]{quantile}}}

\item{omit.reps}{\code{character} vector of conditions for removing
replications.  \code{"no.conv"} removes replications for which the model
did not converge.  \code{"no.se"} removes replications for which
\emph{SE}s could not be calculated.  \code{"no.hpd"} removes replications
with improper solutions (a nonpositive definite matrix).  \code{integer}s
can also be included in the vector, indicating particular replications
to remove.}
}
\value{
\item{show}{\code{signature(object = "FitProp")}: Prints a message
    about data-generation options and how many models converged.}
  \item{summary}{\code{signature(object = "FitProp", ...,
    fit.measures = "srmr", lower.tail = TRUE, NML = FALSE, UIF = FALSE,
    probs = seq(0, 1, .1), type = 7, omit.reps = c("no.conv","no.se"))}:
    For any models in \code{...}, quantiles from the empirical CDFs of
    each \code{fit.measures} are provided, along with summary statistics
    for model comparison (e.g., UIF).}
}
\description{
This class extends \code{\linkS4class{lavaanList}} to specialize in fitting
a hypothesized model to random data patterns, thus accounting for model
complexity not merely as \emph{df} or number of estimated parameters, but
as fitting propensity (Bonifay & Cai, 2017; Preacher, 2003, 2006).
}
\section{Slots}{

\describe{
\item{\code{lavaanList_slots}}{All slots are available from
\code{\linkS4class{lavaanList}} via \code{...}}

\item{\code{lavListCall}}{call to \code{\link[lavaan]{lavaanList}} used to fit the
model to the list of random data patterns in \code{@momentList} or
\code{@ranDataList}, stored as a \code{list} of arguments}

\item{\code{convergence}}{\code{list} of \code{logical} vectors indicating whether,
for each random data patterns, (1) the model converged on a solution, (2)
\emph{SE}s could be calculated, (3) the (residual) covariance matrix of
latent variables (\eqn{\Psi}) is non-positive-definite, and (4) the
residual covariance matrix of observed variables (\eqn{\Theta}) is
non-positive-definite.}

\item{\code{custom}}{\code{logical} indicating whether a custom data-generating
function was provided, rather than using a procedure implemented in the
\code{\pkg{ockhamSEM}} package.}

\item{\code{momentList}}{\code{list} of summary statistics for random data patterns}

\item{\code{ranDataList}}{\code{list} of raw data, typically generated from summary
statistics in \code{@momentList} (e.g., \code{if (!@custom)}).}

\item{\code{obsFitMeas}}{\code{\link[lavaan]{fitMeasures}()} output from real data}

\item{\code{fitMeasDist}}{\code{\link[lavaan]{fitMeasures}()} output from each
random data pattern}

\item{\code{lavobjects}}{\code{list} of \code{\linkS4class{lavaan}} objects (always
the \code{$target} model, optionally a \code{$baseline.model}).}
}}

\examples{
## See ?fitprop help page for examples

}
\references{
Bonifay, W. E., & Cai, L. (2017). On the complexity of item response theory
models. \emph{Multivariate Behavioral Research, 52}(4), 465--484.
\doi{10.1080/00273171.2017.1309262}

Preacher, K. J. (2003).
\emph{The role of model complexity in the evaluation of structural equation models}
 (Unpublished doctoral dissertation). The Ohio State University:
\url{http://rave.ohiolink.edu/etdc/view?acc_num=osu1054130634}

Preacher, K. J. (2006). Quantifying parsimony in structural equation modeling.
\emph{Multivariate Behavioral Research, 41}(3), 227--259.
\doi{10.1207/s15327906mbr4103_1}
}
\author{
Terrence D. Jorgensen (University of Amsterdam; \email{TJorgensen314@gmail.com})
}
