% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitoutput.R
\name{intersect.fitprop}
\alias{intersect.fitprop}
\title{Overview of intersections of fitting propensity of included models}
\usage{
intersect.fitprop(
  x,
  whichmod = NULL,
  whichfit = colnames(x$fit_list[[1]]),
  samereps = TRUE,
  cutoff = NULL,
  lower.tail = NULL,
  mod.lab = NULL,
  saveTable = FALSE
)
}
\arguments{
\item{x}{Object of class fitprop, as created by run.fitprop function.}

\item{whichmod}{Index number corresponding to which model(s) to include on the plot. Defaults to all models.}

\item{whichfit}{Character vector indicating which indices of model fit to include. Defaults to all saved indices.}

\item{samereps}{Logical value indicating whether to use only results from replications in which all selected models yielded results.}

\item{cutoff}{Optional numeric vector indicating what cut value of the fit indice(s) to use. When not specified, no cutoff line is included in resulting plots.}

\item{lower.tail}{Logical vector indicating whether lower values of each fit index corresponds to good fit.}

\item{mod.lab}{Optional character vector of labels for each model.}

\item{saveTable}{Logical value indicating whether to save tables to a list (TRUE) or just produce table to output.}
}
\value{
A list of matrices listing the different intersections, their frequency, and proportion within the dataspace.
}
\description{
Overview of intersections of fitting propensity of included models
}
\examples{
\donttest{
#' # Set up a covariance matrix to fit models to
p<-3 # number of variables
temp_mat <- diag(p) # identity matrix
colnames(temp_mat) <- rownames(temp_mat) <- paste0("V", seq(1, p))

# Define and fit two models using lavaan package
mod1a <- 'V3 ~ V1 + V2
  V1 ~~ 0*V2'
mod2a <- 'V3 ~ V1
  V2 ~ V3'

mod1a.fit <- sem(mod1a, sample.cov=temp_mat, sample.nobs=500)
mod2a.fit <- sem(mod2a, sample.cov=temp_mat, sample.nobs=500)

# Run fit propensity analysis (50 reps to save time)
# Onion approach, save srmr and CFI
res <- run.fitprop(mod1a.fit, mod2a.fit, fit.measure=c("srmr","cfi"),
  rmethod="onion",reps=50)

# Generate intersection tables
intersect.fitprop(res)

# Specify custom cutoffs and lower.tail
intersect.fitprop(res, cutoff = c(.08, .95), lower.tail = c(TRUE, FALSE))

# Save tables as a list and access first one.
tables <- intersect.fitprop(res, saveTable = TRUE)
tables[[1]]

}

}
