% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitoutput.R
\name{pairwise.fitprop}
\alias{pairwise.fitprop}
\title{Plot of pairwise fitting propensity for a specific fit index across possible values}
\usage{
pairwise.fitprop(
  x,
  m = m,
  fm = fm,
  whichmod = NULL,
  savePlot = FALSE,
  xlim = c(0, 1),
  samereps = TRUE,
  lower.tail = NULL,
  mod.lab = NULL,
  mod.brewer.pal = "Set1"
)
}
\arguments{
\item{x}{Data frame of specific fit index values extracted from object of class fitprop, as created by run.fitprop function.}

\item{m}{Integer value indicating location of fit index in list of fit indices to plot}

\item{fm}{Character value indicating which fit index to plot}

\item{whichmod}{Index number corresponding to which model(s) to include on the plot. Defaults to all models.}

\item{savePlot}{Logical value indicating whether to save plot to a list (TRUE) or just produce plot to output.}

\item{xlim}{Numeric vector of length 2 indicating the limits for the x-axis of the plot.}

\item{samereps}{Logical value indicating whether to use only results from replications in which all selected models yielded results.}

\item{lower.tail}{Logical vector indicating whether lower values of each fit index corresponds to good fit.}

\item{mod.lab}{Optional character vector of labels for each model.}

\item{mod.brewer.pal}{Optional string corresponding to the palette from RColorBrewer to use for the different models. e.g.,
see \code{\link[RColorBrewer]{display.brewer.all}}.}
}
\value{
A ggplot object of the pairwise FP comparison plot
}
\description{
Plot of pairwise fitting propensity for a specific fit index across possible values
}
