% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_prop.R
\name{plot.fitprop}
\alias{plot.fitprop}
\title{Plot function for fitprop objects}
\usage{
\method{plot}{fitprop}(
  x,
  ...,
  type = "ecdf",
  whichmod = NULL,
  whichfit = colnames(x$fit_list[[1]]),
  savePlot = FALSE,
  xlim = c(0, 1),
  samereps = TRUE,
  cutoff = rep(0.1, length(whichfit)),
  lower.tail = rep(TRUE, length(whichfit)),
  mod.lab = NULL,
  mod.brewer.pal = "Set1"
)
}
\arguments{
\item{x}{Object of class fitprop, as created by fit_prop function.}

\item{...}{Does nothing but to hopefully make this generic function pass R CMD check}

\item{type}{What type of plot to produce?}

\item{whichmod}{Index number corresponding to which model(s) to include on the plot}

\item{whichfit}{Character vector indicating which indices of model fit to include}

\item{savePlot}{Logical value indicating whether to save plot to a list (TRUE) or just produce plot to output.}

\item{xlim}{Vector of length 2 indicating the limits for the x-axis of the plot}

\item{samereps}{Logical value indicating whether to use only results from replications in which all selected models yielded results}

\item{cutoff}{Numeric vector indicating what cut value of the fit indice(s) to use for euler plots}

\item{lower.tail}{Logical vector indicating whether lower values of each fit index corresponds to good fit}

\item{mod.lab}{Optional character vector of labels for each model}

\item{mod.brewer.pal}{Optional character corresponding to the palette from RColorBrewer to use for the different models}
}
\description{
Plot function for fitprop objects
}
