% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitprop.R
\name{run.fitprop}
\alias{run.fitprop}
\title{Run fit propensity analyses}
\usage{
run.fitprop(
  ...,
  fit.measure = c("srmr"),
  saveModel = FALSE,
  saveR = FALSE,
  control = list(eta = 1, reps = 1000, rmethod = "onion", onlypos = FALSE, seed = 1234),
  cluster = NULL
)
}
\arguments{
\item{...}{models to be fit to the same data of class lavaan}

\item{fit.measure}{character string or vector that indicates which fit measure to extract from fitted models}

\item{saveModel}{logical value indicating whether the to save fitted models for later examination}

\item{saveR}{logical value indicating whether to save random correlation matrices}

\item{control}{list of parameters for generation of random correlation matrices}

\item{cluster}{a cluster created by makeCluster() from the parallel package}
}
\value{
An object of class fitprop
}
\description{
Run fit propensity analyses
}
