% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_prop.R
\name{summary.fitprop}
\alias{summary.fitprop}
\title{Summary function for fitprop objects}
\usage{
\method{summary}{fitprop}(
  object,
  ...,
  probs = seq(0, 1, 0.1),
  samereps = TRUE,
  lower.tail = rep(TRUE, ncol(object$fit_list[[1]]))
)
}
\arguments{
\item{object}{Object of class fitprop, as created by fit_prop function.}

\item{...}{Does nothing but to hopefully make this generic function pass R CMD check}

\item{probs}{passed to quantile to determine what probabilities to report}

\item{samereps}{Logical value indicating whether to use only results from replications in which all selected models yielded results}

\item{lower.tail}{Logical vector indicating whether lower values of each fit index corresponds to good fit}
}
\description{
Summary function for fitprop objects
}
